# Refactor Logger