--- a/src/com/config/CommunicationConfiguration.hpp
+++ b/src/com/config/CommunicationConfiguration.hpp
@@ -2,6 +2,7 @@
 
 #include "logging/Logger.hpp"
 #include "xml/XMLTag.hpp"
+#include "com/SharedPointer.hpp"
 
 namespace precice
 {