--- a/src/logging/Logger.cpp
+++ b/src/logging/Logger.cpp
@@ -50,6 +50,8 @@ Logger::Logger(const Logger& other)
 {
 }
 
+Logger::Logger(Logger&& other) = default;
+
 Logger& Logger::operator=(Logger other)
 {
     swap(other);