--- a/src/logging/Logger.hpp
+++ b/src/logging/Logger.hpp
@@ -6,29 +6,39 @@
 namespace precice {
 namespace logging {
 
+/// Struct used to capture the original location of a log request
 struct LogLocation {
     const char * file;
     int line;
     const char * func;
 };
 
+/// This class provides a leightweight logger.
 class Logger {
 public:
+  /** creates a logger for a given module.
+   * @param[in] the name of the module 
+   */
   explicit Logger(std::string module);
+
   Logger(const Logger& other);
   Logger& operator=(Logger other);
   ~Logger();
 
   void swap(Logger& other) noexcept;
 
+  ///@name Logging operations
+  ///@{
   void error(LogLocation loc, const std::string& mess);
   void warning(LogLocation loc, const std::string& mess);
   void info(LogLocation loc, const std::string& mess);
   void debug(LogLocation loc, const std::string& mess);
   void trace(LogLocation loc, const std::string& mess);
-
+  ///@}
 private:
+  /// Forward declaration of the implementation of the logger
   class LoggerImpl;
+  /// Pimpl to the logger implementation
   std::unique_ptr<LoggerImpl> _impl;
 };