--- a/src/logging/Tracer.cpp
+++ b/src/logging/Tracer.cpp
@@ -7,28 +7,16 @@ namespace logging {
 Tracer::Tracer
 (
   Logger &log,
-  std::string function,
-  std::string file,
-  long line
+  LogLocation loc
   )
   :
   _log(log),
-  _function(function),
-  _file(file),
-  _line(line)
+  _loc(std::move(loc))
 {}
 
 Tracer::~Tracer()
 {
-  using namespace boost::log;
-
-  attribute_cast<attributes::mutable_constant<int>>(core::get()->get_global_attributes()["Line"]).set(_line);
-
-  attribute_cast<attributes::mutable_constant<std::string>>(core::get()->get_global_attributes()["File"]).set(_file);
-
-  attribute_cast<attributes::mutable_constant<std::string>>(core::get()->get_global_attributes()["Function"]).set(_function);
-
-  BOOST_LOG_SEV(_log, trivial::severity_level::trace) << "Leaving " << _function;
+  _log.trace(_loc, std::string{"Leaving "}.append(_loc.func));
 }
 
 }} // namespace precice,logging