--- a/src/utils/String.hpp
+++ b/src/utils/String.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <sstream>
 
 namespace precice {
 namespace utils {
@@ -23,6 +24,13 @@ std::string & checkAppendExtension(std::string& filename, const std::string& ext
  * This function is case-insensitive.
  */
 bool convertStringToBool(std::string const & value);
+
+
+#define PRECICE_AS_STRING(message) [&]{ \
+        std::ostringstream oss;        \
+        oss << message;                \
+        return oss.str();              \
+    }()
 
 
 }} // namespace precice, utils