This PR refactors the current logger.

## Problem:
* Due to the Boost dependencies, the cost of including `Logger.hpp` is around 2.6 seconds, including:
   * 1500ms for `boost/log/expressions.hpp`via `LogMacros.hpp`
   * 810ms for `boost/log/trivial.hpp` aka the actual logger
   * ~250ms for `MasterSlave.hpp`
   * 100ms for `prettyprint.hpp` (Depends strongly on the call-site)
* This header is included in nearly every translation unit, which slows the entire build process down.

## My Approach:
* Use a pimpl to move the boost dependencies into a single translation unit `Logger.cpp`
* Use a minimal struct to forward the log location to the impl.
* Materialize passed log message to a string and pass that to the impl.
* Use functions for the log levels to prevent a boost dependency for the log-level in the interface.

## Side-Effects:
* The Logger itself is now a mere pointer, thus doesn't bloat up the class it is member of.
* Accessing the logger is always an indirection.

## Result:
* The cost of including `Logger.hpp` now converges against `prettyprint.hpp`.
* The build runs notably faster (Debug + `make -j4` on i7-7700 in tempfs):

|  | Real | User | Sys |
| --- | --- | --- | --- |
Before | 3m43s |  11m29s | 0m37s
After | 2m6s | 6m25s | 0m20s