@fsimonis 2019-03-30
@Summerdave 
> std::ostream /* or whatever BOOST_LOG... returns */ Logger::debug(LogLocation loc){

That was my initial thought too. The [documentation](https://www.boost.org/doc/libs/1_65_0/libs/log/doc/html/log/detailed/sources.html#log.detailed.sources.severity_level_logger) shows an example of how to do something similar.

This however requires to return a proxy object, which holds an opened record and pushes the record on destruction. This proxy has to inherit from `std::ostream` and needs to be returned as a value, meaning the proxy needs to be a known type. To prevent boost in the header, I would have to create a pimpl.  _it goes on and on and I haven't reached the end yet_

The materialization happens somewhere anyhow and it is a minimally invasive and very simple solution.